Add cart functionality to ecommerce API
Implemented a cart functionality for the ecommerce API. This includes routes, controller functions, and model updates. The cart module now allows creating a cart, adding products to the cart, viewing cart contents, and checking out. On the models side, added 'onDelete: CASCADE' for the 'userId' foreign key in the Cart model. Also improved the associations in the Cart model, introducing aliases for more readable code and added indexes for better performance.