Add SSL and connection test to Sequelize config
The Sequelize configuration now includes SSL dialect options and a connection test. SSL for db connection is made mandatory and a note of caution has been added for proper SSL configuration on production environments. Additionally, a test connection function is added which attempts to connect to the database and logs the outcome.