Implement cartId parameter in createOrder method
The createOrder function in the orderController has been modified to accept an optional 'cartId' parameter from the client's request. With this change, a specific cart can be specified when creating an order, instead of automatically selecting the user's cart. This adds flexibility and allows greater control over the order creation process for the clients. The code for querying the CartModel has been updated to accommodate this change.