Enable CORS whitelist and handle origin permission
The commented-out whitelist of approved origins has been activated. The CORS configuration has been modified as well, now it checks if the incoming origin matches any address in the whitelist or if the origin does not exist. A callback is then issued depending on whether the CORS request is approved or rejected.