Refactor user controller and update model associations
Changes in the user controller include using the findByEmail function for user registration and including the first and last names in the user object. The dotenv required in the config file has been updated for a more straightforward path. In the models file, commented-out code has been removed for clarity and readability, letting the active association definitions stand out.